backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.3
python-dotenv==1.0.1
pymongo==4.6.1
pillow==10.2.0
pytesseract==0.3.10
easyocr==1.7.1
torch==2.0.0
redis==4.6.0
requests==2.31.0
numpy<2
python-multipart==0.0.9